backend/requirements.txt
httpx
python-dotenv
python-multipart
fastapi
uvicorn[standard]
openai>=1.1.0
